amend readme
