--- index.js
@@ -1,5 +1,0 @@
-"use strict";
-
-exports.propPaths = require("./dist/nested-prop-paths").propPaths;
-exports.getVals = require("./dist/nested-prop-paths").getVals;
-exports.find = require("./dist/nested-prop-paths").find;
