+++ package.json
@@ -3,17 +3,21 @@
   "version": "0.5.0",
   "homepage": "https://github.com/HBDunn/nested-prop-paths",
   "description": "find all paths/values to a key/prop in array of objects or object",
-  "main": "./dist/nested-prop-paths.js",
-  "browser": "./lib/nested-prop-paths.js",
-  "module": "./module/nested-prop-pathsjs",
+  "main": "./dist/index.js",
+  "browser": "./lib/index.js",
+  "module": "./module/index.js",
   "scripts": {
     "test": "BABEL_ENV=jest jest ",
-    "build": "babel ./src -o ./dist/nested-prop-paths.js --source-maps --copy-files",
-    "build-browser": "BABEL_ENV=browser babel ./src -d ./lib --source-maps --copy-files",
+    "build": "babel ./src -d ./dist --source-maps --copy-files",
+    "build-browser": "BABEL_ENV=browser babel ./src -d ./lib/ --source-maps --copy-files",
     "build-module": "BABEL_ENV=module babel ./src -d ./module --source-maps --copy-files",
-    "build-ex": "babel ./example/try.js -o ./dist/try.js --source-maps --copy-files",
-    "build-all ": "npm run build && npm run build-ex && npm run build-browser && npm run build-module"
+    "clean_dist":"rm -r ./dist &> /dev/null",
+    "clean_lib":"rm -r ./lib &> /dev/null",
+    "clean_module":"rm -r ./module &> /dev/null",
+    "clean": " npm run clean_dist && npm run clean_lib && npm run clean_module",
+    "build-all":  "npm run clean && npm run build && npm run build-browser && npm run build-module"
   },
+  "pre-commit": "build-all",
   "keywords": [
     "json",
     "data",
@@ -24,12 +28,10 @@
     "array"
   ],
   "author": "hdunn <hdunn.id@gmail.com>",
-  
   "repository": {
-    "type" : "git",
-    "url" : "https://github.com/HBDunn/nested-prop-paths"
+    "type": "git",
+    "url": "https://github.com/HBDunn/nested-prop-paths"
   },
-  
   "license": "MIT",
   "devDependencies": {
     "@babel/cli": "7.8.4",
@@ -38,6 +40,8 @@
     "babel-eslint": "^10.1.0",
     "eslint": "6.8.0",
     "eslint-plugin-jest": "^23.8.2",
-    "jest": "25.1.0"
+    "jest": "25.1.0",
+    "pre-commit": "^1.2.2",
+    "travis-cli": "^1.0.9"
   }
 }
