+++ src/index.js
@@ -1,0 +1,6 @@
+"use strict";
+
+exports.propPaths = require("./nested-prop-paths").propPaths;
+exports.getVals = require("./nested-prop-paths").getVals;
+exports.find = require("./nested-prop-paths").find;
+exports.runExample = require("./example/try").runExample;
